Implement a few tricks learned from souper
* Forbid old programs that already failed verification.
* Allow synthesizing programs of a specific length, so we can find the smallest
  programs.
* Have the solver choose the initial inputs for finite synthesis.

+++ src/builder.rs
@@ -2,7 +2,7 @@ use crate::{Id, Instruction, Operator, Program};
 
 #[derive(Debug)]
 pub struct ProgramBuilder {
-    id_counter: usize,
+    id_counter: u32,
     program: Program,
 }
